Контроллеры отредактированы. Проблема с showUser не решена.
