--- src/main/java/ru/morrigan/spring/boot_security/demo/controller/LoginController.java
@@ -1,19 +1,0 @@
-package ru.morrigan.spring.boot_security.demo.controller;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-
-
-@Controller
-public class LoginController {
-    @GetMapping("/")
-    public String index() {
-        return "login";
-    }
-
-    @RequestMapping(value = "/login", method = RequestMethod.GET)
-    public String loginPage() {
-        return "login";
-    }
-}
