+++ src/main/java/ru/morrigan/spring/boot_security/demo/controller/UserController.java
@@ -8,14 +8,13 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import ru.morrigan.spring.boot_security.demo.service.UserService;
 
 @AllArgsConstructor
 @RequestMapping("/")
 @Controller
 public class UserController {
-    @Autowired
-    private UserService userService;
 
     @GetMapping("/user")
     public String userInfo(Model model) {
@@ -23,5 +22,15 @@ public class UserController {
                 SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         model.addAttribute("user", user);
         return "showUser";
+    }
+// Перенос из Логин контроллера
+    @GetMapping("/")
+    public String index() {
+        return "login";
+    }
+
+    @RequestMapping(value = "/login", method = RequestMethod.GET)
+    public String loginPage() {
+        return "login";
     }
 }
